refactor(seats): update seat selection logic
Refactored the seat selection logic in the Seats component to improve readability
and maintainability. Added a new custom hook `useSeatSelection` to handle seat
selection based on rows and columns. Cleaned up the grouping of seats by row
and column to simplify the code and make it more efficient.
