+++ src/lib/seat.ts
@@ -10,3 +10,38 @@ export interface Seat {
 	indexFromLeft: number;
 	status: SeatStatus;
 }
+
+export function getSeatsByRow(seats: Seat[]): Record<number, Seat[]> {
+	const seatsByRow = seats.reduce<Record<number, Seat[]>>((acc, seat) => {
+		const { row } = seat;
+		// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition --- its fine
+		acc[row] = acc[row] || [];
+		acc[row].push(seat);
+		return acc;
+	}, {});
+
+	return seatsByRow;
+}
+
+export function getSeatsByColumn(
+	seatsByRow: Record<number, Seat[]>,
+): Record<number, Record<number, Seat[]>> {
+	const seatsByColumn: Record<number, Record<number, Seat[]>> = Object.keys(
+		seatsByRow,
+	).reduce<Record<number, Record<number, Seat[]>>>((acc, rowKey) => {
+		const rowSeats = seatsByRow[Number(rowKey)];
+		acc[Number(rowKey)] = rowSeats.reduce<Record<number, Seat[]>>(
+			(rowAcc, seat) => {
+				const { column } = seat;
+				// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition --- its fine
+				rowAcc[column] = rowAcc[column] || [];
+				rowAcc[column].push(seat);
+				return rowAcc;
+			},
+			{},
+		);
+		return acc;
+	}, {});
+
+	return seatsByColumn;
+}
